# feat: Add comprehensive Frame functionality with Miro-inspired adoption system