… .claude/settings.local.json
… README.md
… app/board/[boardId]/_components/toolbar.tsx
… docs/FRAME_COMPONENT_DESIGN.md
… docs/PR_FRAME_FUNCTIONALITY.md
… docs/architecture/FRAME_FEATURE_DESIGN.md
… hooks/use-layer-editing.ts
… lib/utils.ts
… stores/canvas-store.ts
… types/canvas.ts
… types/profile.ts